Moved internal functions
